assistant/language_model/requirements.txt
openai
python-dotenv
dirigera
